backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-dotenv
slowapi
python-multipart
aiosqlite
google-generativeai
